refactor(server): move external MCP validation under agents/ module
The endpoint validates an external MCP server the agent piece will
connect to at flow-execution time — that's the agent-piece domain, not
the Activepieces-as-MCP-server feature. Living under /mcp-server/ made
the two directions look related when they aren't.

- New module packages/server/api/src/app/agents/ with
  agent-tools-controller.ts + mcp-tool-validator.ts
- Endpoint moves from POST /v1/projects/:projectId/mcp-server/validate-agent-mcp-tool
  to POST /v1/projects/:projectId/agent-tools/mcp/validate
  (resource-oriented: agent tools > MCP kind > validate action)
- mcp-server-controller.ts goes back to just the AP-as-MCP-server
  endpoints (get, update, rotate)
- Web client + feature docs follow the new path
