+++ packages/web/src/features/agents/agent-tools/mcp-tool-dialog/api.ts
@@ -11,7 +11,7 @@ export const mcpToolApi = {
     request: AgentMcpTool,
   ): Promise<ValidateAgentMcpToolResponse> {
     return await api.post<ValidateAgentMcpToolResponse>(
-      `/v1/projects/${projectId}/mcp-server/validate-agent-mcp-tool`,
+      `/v1/projects/${projectId}/agent-tools/mcp/validate`,
       request,
     );
   },
